[Add] #17 - UserData struct 분리
