+++ 35-seminar/Entity/UserData.swift
@@ -1,0 +1,15 @@
+//
+//  UserData.swift
+//  35-seminar
+//
+//  Created by 김유림 on 11/11/24.
+//
+
+import Foundation
+
+struct UserData {
+    var username: String
+    var password: String
+    var token: String
+    var autoLogin: Bool
+}
