+++ 35-seminar/Presentation/Appstore/User/View/UserAccountView.swift
@@ -1,0 +1,20 @@
+//
+//  UserAccountView.swift
+//  35-seminar
+//
+//  Created by 김유림 on 11/11/24.
+//
+
+import UIKit
+
+class UserAccountView: UIView {
+
+    /*
+    // Only override draw() if you perform custom drawing.
+    // An empty implementation adversely affects performance during animation.
+    override func draw(_ rect: CGRect) {
+        // Drawing code
+    }
+    */
+
+}
