+++ 35-seminar/Presentation/Appstore/User/ViewController/UserAccountViewController.swift
@@ -1,0 +1,29 @@
+//
+//  UserAccountViewController.swift
+//  35-seminar
+//
+//  Created by 김유림 on 11/11/24.
+//
+
+import UIKit
+
+class UserAccountViewController: UIViewController {
+
+    override func viewDidLoad() {
+        super.viewDidLoad()
+
+        // Do any additional setup after loading the view.
+    }
+    
+
+    /*
+    // MARK: - Navigation
+
+    // In a storyboard-based application, you will often want to do a little preparation before navigation
+    override func prepare(for segue: UIStoryboardSegue, sender: Any?) {
+        // Get the new view controller using segue.destination.
+        // Pass the selected object to the new view controller.
+    }
+    */
+
+}
